[TECH-496] Rename mpyl.py to build.py
branch: feature/TECH-496-improve-code-quality
